+++ src/models/webhook.js
@@ -18,11 +18,6 @@ module.exports = function ({ samples }) {
       { key: 'createdAt', label: 'Created At', type: 'date' },
       { key: 'updatedAt', label: 'Updated At', type: 'date' },
     ],
-    filters: {
-      workspaceId: {
-        required: true,
-        type: 'number',
-      },
-    },
+    filters: {},
   }
 }