+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+# 1.0.14
+
+- Remove deprecated filters
+- Add information about default values of filters
+
 # 1.0.13
 
 - Add changelog